refactor: move showToast to utils
