+++ src/redux/downloads/downloads.actions.js
@@ -2,7 +2,7 @@ import { CANCEL_DOWNLOAD, PAUSE_DOWNLOADS } from './donwloads.types';
 import BackgroundService from 'react-native-background-actions';
 import { downloadChapter } from '@database/queries/ChapterQueries';
 import { CHAPTER_DOWNLOADED } from '../novel/novel.types';
-import { showToast } from '@hooks/showToast';
+import { showToast } from '@utils/showToast';
 
 import * as Notifications from 'expo-notifications';
